+++ nodes/Autotask/resources/companies/execute.ts
@@ -8,6 +8,7 @@ import {
 	CountOperation,
 } from '../../operations/base';
 import { searchCompaniesByDomain } from '../../helpers/company-domain-search';
+import { getSelectedColumns } from '../../operations/common/select-columns/filter-entity';
 
 const ENTITY_TYPE = 'company';
 
@@ -68,11 +69,13 @@ export async function executeCompanyOperation(
 					const domainOperator = this.getNodeParameter('domainOperator', i, 'contains') as string;
 					const searchContactEmails = this.getNodeParameter('searchContactEmails', i, true) as boolean;
 					const limit = this.getNodeParameter('limit', i, 25) as number;
+					const selectColumns = getSelectedColumns(this, i);
 					const response = await searchCompaniesByDomain(this, {
 						domain,
 						domainOperator,
 						searchContactEmails,
 						limit,
+						selectColumns,
 						itemIndex: i,
 					});
 					returnData.push({ json: response });